Fix click outside
